fix : 회원정보 업데이트 이미지 수정 #73

+++ src/main/java/com/project/date/dto/response/MypageResponseDto.java
@@ -15,7 +15,7 @@ public class MypageResponseDto{
     private Long profileId;
     private String nickname;
     private String imageUrl;
-    private Long age;
+    private Integer age;
     private String mbti;
     private String introduction;
     private String area;
